requirements.txt
pymsteams
azure.cli.core
azure-cli
msrestazure
packaging
